[COM_RPC] Fix a crash when calling SHOW SLAVE STATUS while using a detached session
Summary: The code in rpl_slave.cc that was run when SHOW SLAVE STATUS was executed accessed the thd's packet buffer directly.  Since this was a detached session we want it to use the information in the connection thd's packet buffer.  This was available via the protocol, so I switched over to access the packet (and net) via the protocol object.  Also added a test that ran a bunch of SHOW commands to make sure none of the other ones had a similar problem.

Test Plan: MTR with new main.com_rpc_show test

Reviewers: jprahman, herman, anca, mung, #dbclient

Reviewed By: jprahman

Subscribers: abalsara, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D7783389

Tasks: T28404312

Signature: 7783389:1524809813:0289d1dcad67ef41cf6d09b63814d2db7b5ed090