Fix crash in calculate_stats() when tables are modified/removed
Summary:
During periodic stats recalculation, the keydefs of tables to
recalculate are pulled out into the to_recalc array. This is passed into
calculate_stats where it is used to find the set of sst files containing
the indexes. Within calculate_stats(), there is a second call to find
the keydef from the ddl_manager, but it is possible the table has been
removed when the second call is made. The code does not check the
contents of the keydef pointer before de-referencing it. If the table has
changed, the pointer is NULL, and can cause a crash.

Since the key definition is already being passed in, it can be reused
instead of another lookup in the ddl_manager. To make this more
efficient, change to_recalc to a map of gl_index_id -> key_def, so that
lookups can be performed on to_recalc for the stored key_def.

Squash with D7482233

Test Plan:
Setup test instance and run myrocks. Repeatedly call analyze table and
set rocksdb_stats_recalc_rate to a high value

Reviewers: mung, rtm

Reviewed By: mung

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D7920320

Tasks: T29132315

Signature: 7920320:1525822965:52f50f708eaa800d209186535ec71410645b8464