[COM_RPC] Fix "Error 1 during COMMIT" error
Summary:
The query_id for the detached session's THD was never getting set correctly.  This is the piece of information that is included in the RocksDB transaction to make it unique.  Updated the code to copy it from the connection's THD.

The new test failed every time with the "Error 1 during COMMIT" error and succeeds with the fix in place.

Test Plan: MTR with new rocksdb.com_rpc_tx test

Reviewers: herman, aasogamo, anca, mung, #dbclient

Reviewed By: aasogamo

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D8013383

Tasks: T29354371

Signature: 8013383:1526423250:1d87c390dcd3b6692808cc432029adf6dd8b5dc6