Fix MyRocks SK updates when master is missing columns
Summary:
When master has missing columns in the middle of the table and COMPLETE binlog
format is used SKs were not updated on properly on the slave. This was
happening because a temporary code was resetting the tables's write set after
it was set correctly using column names and MyRocks uses the write set to
compute SKs that need to be updated.

Removed the temporary code to fix the issue.

Test Plan: MTR

Reviewers: herman, yoshinori

Reviewed By: herman

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D8457633

Tasks: T30047854

Signature: 8457633:1529107966:65bb0454e2a8ef2d66834e5e7d3543f5d4abd42b