Fix writing binlogs with checksums during slave recovery
Summary:
During slave recovery (both slave_gtid_info and idempotent based) events for
which the changes are already applied in the engine are written to the binlogs
directly without touching the engine. These events are written to the binlogs
though the binlog cache. If checksums are enabled, they are generated for each
event just before flushing the binlog cache to disk (in
`MYSQL_BIN_LOG::do_write_cache()`).Now, during recovery when we write an event
to the cache we might end up writing it along with the checksum (if they were
generated by the master), then during binlog cache flush we'll calculate a
checksum again. This might causes an extra checksum to be written which causes
errors during subsequent parsing.

In this change we modify the `Log_event::need_checksum()` method and skip
writing checksums for events which are being written directly to the binlogs
(skipping the storage engine). This prevents double writing of checksums.

Test Plan: MTR

Reviewers: herman, yoshinori, arahut

Reviewed By: herman

Subscribers: jjj, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D8672520

Tasks: T30416647

Signature: 8672520:1530814356:f92fc044f6f92294a8ffa12d75800e9443169377