[db_metadata]: Fix for corruption of existing value after alter without value is executed
Summary:
The db.opt file is re-written every time an ALTER DATABASE statement is ran. In case an ALTER is ran without changing the DB_METADATA for a database, then the DB_METADATA may get corrupted.
This is because of a bug in the code: If db_metadata is not specified in the ALTER statement, `thd->lex->create_info->db_metadata` has garbage value. Before re-writing the db.opt file, the existing (previous) options are loaded from the hash (or db.opt file) and any option that is not specified in the current ALTER statement being executed is defaulted to the existing value. This was not done for the db_metadata option. Hence instead of using the existing value by default, the garbage value was used and written to the file and the hash.

This diff fixes the above bug.

Test Plan: Added unit tests in the existing test

Reviewers: abhinavsharma

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D9445354

Tasks: T32631803