[binlog_meta_data] Fixed timestamp extraction
Summary:
The JSON key was changes from "timestamps" to "ts" to save space in
binlogs. We forgot to make the key change in extraction function.

Test Plan:
Added a debug assert that would fail in the earlier version but pass
in the correct version. Ran MTR test with the assert.

Reviewers: yashb

Reviewed By: yashb

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D9544555

Signature: 9544555:1535484313:21a869cc166101121ddb0f81e814a71ff072176f