[semisync_whitelist] Fixed valgrind issues
Summary:
When testing if a invalid semi-sync acker's dump thread is killed when
an ACK is recieved from it, we were not making sure that the ACK
received is from the invalid slave. So, sometimes in lengthy valgrind
runs we would send the transaction to the invalid slave after the valid
slave ACKed the transaction. Since, any transaction sent to semi-sync
slaves after an ACK will not have the magic header set, the invalid
slave will never send an ACK. So the dump thread never dies and we
timeout in our test while waiting for it to die.

Fixed the test by stopping the valid slave so we receive ACK from invalid one.

Test Plan: Ran full-tests twice and saw that the test passed both times.

Reviewers: rtm

Reviewed By: rtm

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D9551467

Tasks: T32875050

Signature: 9551467:1535567584:3741676f170817dd91e238e3ff2170f1549771ec