Stabilize main.com_rpc test
Summary:
The main.com_rpc test is flakey sometimes because during the kill test, it is possible that the killed query hasn't exited before we check for its status on processlist. This is because kill is asynchronously run.

The fix is to add a wait.

Test Plan: mtr

Reviewers: jkedgar, abalsara

Reviewed By: jkedgar

Subscribers: webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D9975738

Tasks: T20952138

Signature: 9975738:1537467323:2fb4e4a692cef0bae7e40fac49cba2c2d0689b8c