Extend optimizer_low_limit_heuristic to skip low limit heuristic everywhere
Summary:
For some reason, this heuristic is applied in two places. One on
a per-table basis in make_join_select, and another one on a global basis at
the end of JOIN::optimize (this one just looks at the first JOIN_TAB).

The second one is actually the more common occurrence, and was not considered
in the original implementation of this variable.

This applies the check in the more common path by assuming that adding a limit
never makes an index scan cheaper. The rationale is that it's hard to estimate
how much cheaper a plan is with limit, and a bad estimate can lead to full
index scans.

Squash with: D5187687

Test Plan: mtr

Reviewers: dmituzas, herman

Reviewed By: herman

Subscribers: butterflybot, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D13148255

Signature: 13148255:1543366861:2f86176b6fa1ae73f40aa7c15cdea995e8ca9551