Add high priority dll mtr test for task T33221077
Summary: This change contains the high priority ddl test for task T33221077. High priority dll is a FB internal feature and so this test is not included in the original diff, which is going to be submitted to ORACLE.

Test Plan: The change itself is a mtr test. The test uses high priority dll feature to kill slave sql worker thread. The test repros the crash before the fix and passes after the fix.

Reviewers: herman, mung, rtm, vasilep

Reviewed By: herman

Subscribers: butterflybot, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D13457834

Tasks: T33221077

Signature: 13457834:1544823314:6b879203af48959783391e4f54e0d7ea2380c234