Remove the RPC_ID from the info portion of the ok packet
Summary: At one point we returned the rpc_id in the info portion of the info packet.  Once we switched to using response attributes we needed to leave it in until MyRouter looked for the data in the response attributes.  But then it was forgotten.  Now when turning on rpc_protocol, it causes a problem in the main.alter_table test as that test turns on the display of the info field.  Remove the rpc_id from the info portion.

Test Plan: mysqltest.sh --rpc_protocol main.alter_table

Reviewers: herman, mung, abalsara, #dbaas

Reviewed By: abalsara

Subscribers: butterflybot, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D14313578

Tasks: T29491753

Signature: 14313578:1551809270:ae01b1fcf79eb891ca4f16c742e5eb4b6ba903b1