[dependency replication] Sync transaction when number of keys exceed a threshold
Summary:
When a transaction modifies too many rows the map which is used to hold
keys (to find dependencies) can become too large. This results in slow lookups
and memory pressure. This diff introduces a new variable mts_dependency_max_keys
(default: 100000) that defines a threshold after which the transaction is
executed in isolation (i.e. we wait for all transactions before it to complete
and then execute it isolation). Since the transaction is executed in isolation,
its key information is not stored

Also changed the hash function for keys to well known murmur hash which has a
better distribution.

Test Plan: MTR

Reviewers: arahut

Reviewed By: arahut

Subscribers: butterflybot, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D14235731

Signature: 14235731:1554358045:1846bd6835ee728af22777cbcde4b26da21e9c90