mysqlslap honors protocol
Summary: I find out in my testing that mysqlslap doesn't really use the --protocol parameter in its workers. The bigger problem seems to be that it doesn't apply some of its parameters in the worker thread, --protocol is just one of them. I'm not sure I have time to test all the parameters here so I'm just adding the fix for --protocol. The testing for mysqlslap seems be rather minimal in MySQL in general.

Test Plan: Manual testing. I'll see if I can add a test for it, but it may end up not worth in the context of this two line fix.

Reviewers: luqun, herman, bintong, mung

Reviewed By: herman

Subscribers: butterflybot, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D15010933

Signature: 15010933:1556152259:cb4474ee350459019ea055285af34f6c92de59d1