Fix binlog corruption during slave recovery when master-slave schema don't match
Summary:
During recovery in multi-threaded slave (both idempotent and
slave_gtid_info styles) to fill the gaps in binlog some rows events
blindly copied from the relay logs to the binlogs. If the schema is
different between the master and slave, this causes a situation where
for a recovered transaction the table map event reflects the slave's
schema but the rows event reflects the master's schema. Anything that
parses this transaction in the slave's binlog will error out due to this
mismatch because it won't be able to parse the rows event using the
table map event.

To solve this issue now we massage the rows event to reflect the slave's
schema (basically by unpacking it and then packing it again) before
writing it.

Test Plan: MTR

Reviewers: herman, yoshinori, arahut

Reviewed By: herman

Subscribers: butterflybot, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D14944825

Tasks: T43878722, T43838220, T42545312

Signature: 14944825:1556732596:354910e8bedda9ff60534a3af437c55035c1863a