[MyRocks8.0] Use unordered_map instead of my_hash
Summary:
This is part of MyRocks 8.0 prep work - front loading as much work as possible in 5.6 before we actually port MyRocks to 8.0.

In MySQL 8.0 the my_hash completely stuff is gone and replaced by unordered_map and various sub-class (such as collation_unordered_map, etc). And in their source code blog they claim that by switching to unordered_map gives them performance wins. So it makes sense for us to switch to it preemptively before we port to MySQL 8.0.

This is rewrite/enhancements of below changes from georgelorchpercona (not a literal port):

https://github.com/facebook/mysql-5.6/pull/930/commits/20d4b2b7376194144ce05ef02949ea1f479928c5
https://github.com/facebook/mysql-5.6/pull/930/commits/8a5e758dbdd5c4b9e4f5839bb0791edfac4e7638

Test Plan:
1. Sandcastle tests
2. MyShadow to ensure no perf regression
3. Validate OOM behavior (though I kinda suspect the answer is just that MySQL will crash)

Reviewers: luqun, herman, bintong, mung

Reviewed By: mung

Subscribers: butterflybot, flor, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D15313077

Signature: 15313077:1557864956:d5e336e6ad6918d982443789104a195b89bd0852