[mysql][ hlc] Return hlc timestamp for a write txn
Summary:
The HLC timestamp associated with a GTID of a txn is returned as part of OK
packets. The HLC timestamp is stored in session response attributes tracker
field of the OK packet. The HLC is stored as a key-value pair in session
response attribute - the key is 'hlc_ts' and the value is the HLC timestamp.
Setting the HLC timestamp in OK packets is gated by a sysvar
'session_track_hlc'.

Only a RW transaction (i.e a transaction that genarates a GTID) will have a HLC
timestamp retuned in its OK packet.

Test Plan: New tests and sandcastle

Reviewers: abhinavsharma, jprahman, herman

Reviewed By: jprahman

Subscribers: butterflybot, jprahman, mzait, maiergre, ritwikyadav, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D15987483

Signature: 15987483:1562863074:7be48073b2dcf67c7922480d7ba32783cd2a1d65