Remove caching of m_update_scope data across rows in single update
Summary:
The m_update_scope bitmap is used to determine if a secondary key needs
to be updated. It is calculated for the first row of an update statement
and assumed to be the same for other rows in the update statement.

However, binlogs with minimal or complete images can generate row
changes that have different columns being updated. The different
write_sets per row need to be recalculated for each call to
calc_updated_indexes(), but because the first row's write_set was being
cached, the other rows did not trigger updates for the appropriate
indexes.

Test Plan: mtr

Reviewers: abhinavsharma, mung, yoshinori, yzha

Reviewed By: abhinavsharma

Subscribers: butterflybot, zhichengzhu, ritwikyadav, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D16383492

Tasks: T47585582, T47391619

Signature: 16383492:1563580697:e377360d8d1e3f64d2b82708afbbdf39c9d25a17