[semi_sync] Calling remove_slave() when plugin fails while starting dump thread
Summary:
When we return an error in repl_semi_binlog_dump_start() we should
cleanup properly and call remove_slave(). This bug was causing the number of
semi_sync client to display incorrect values when the connection was failing due
to whitelist.

Test Plan:
Manually tested by setting the whitelist to '' and reconnecting the
slave multiple times, the counter was 0.

Reviewers: herman, yoshinori, jjj, arahut

Reviewed By: herman

Subscribers: butterflybot, ritwikyadav, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D16151734

Tasks: T41276513

Signature: 16151734:1563828740:026122edd3c9341853cb392a0c2559be83929dba