[dependency replication] Fixing an illegal mem ref bug caused by large trxs
Summary:
Dependency replication uses table map events to find dependencies. The
coordinator assumes that the table map event is always available because in
steady state the worker doesn't destroy events of the transaction till it's
committed. This assumption is not true however, the worker will destroy buffered
transactions after it hits a max threshold. To fix this, we track the number of
events per trx in the coordinator and make it a sync trx when it reaches the
threshold. A sync trx is executed in isolation and hence doesn't do any
dependency calculation.

Test Plan: Wrote an mtr test that fails in valgrind without the fix

Reviewers: herman, mung, arahut

Reviewed By: arahut

Subscribers: butterflybot, ritwikyadav, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D16568788

Tasks: T48227365, T48086252, T47967598

Signature: 16568788:1565196684:2efe4d14ceea1ef7f8054619d7a6445977ce863a