Fix corruption caused by CONCAT function
Summary: The CONCAT function has a special case for when the next item to concatenate in is a substring of the current temporary buffer.  However the check for this case is incorrect.  It checks to see if the pointer is greater or equal to the start of the temporary buffer correctly, but when checking on the end it mistakenly includes the byte after the temporary buffer.  If the next string happens to immediately follow the temporary buffer, the algorithm incorrectly enters this section of code and as a result the ouput buffer ends up with a length greater than the allocated length.  Once this happens, the next time the buffer is copied, only the allocated length is copied though the string continues to assume it had the larger length bytes.  This means that random bytes that happened to be in memory at the new location are treated as part of the string - essentially corrupting the result.

Test Plan: This is very difficult to reproduce.  I do have a test case that can reproduce it right now, but any change to the system that causes a different memory layout will cause it to no longer reproduce, so I'm not including the test here as it will not be useful in the future.

Reviewers: jprahman, herman, mung, yoshinori, yzha, abalsara

Reviewed By: yzha

Subscribers: ritwikyadav, vinaybhat, webscalesql-eng@fb.com

Differential Revision: https://phabricator.intern.facebook.com/D17479735

Tasks: T54257092

Signature: 17479735:1568909610:bb734cc327041d07719ac3b6162d40620e191d4d